+++ NEWS.md
@@ -398,6 +398,9 @@ Deprecated or removed
   * `fill!(A::Diagonal, x)` and `fill!(A::AbstractTriangular, x)` have been deprecated
     in favor of `Base.LinAlg.fillslots!(A, x)` ([#24413]).
 
+  * `eye` has been deprecated in favor of `I` and `Matrix` constructors. Please see the
+    deprecation warnings for replacement details ([#24438]).
+
   * Using Bool values directly as indices is now deprecated and will be an error in the future. Convert
     them to `Int` before indexing if you intend to access index `1` for `true` and `0` for `false`.